chore(app): cleanup main files
