+++ .vscode/settings.json
@@ -9,5 +9,7 @@
 	},
 	"[json]": {
 		"editor.defaultFormatter": "biomejs.biome"
-	}
+	},
+	"rust-analyzer.linkedProjects": ["./src-tauri/Cargo.toml"],
+	"typescript.tsdk": "node_modules/typescript/lib"
 }
