+++ src/eventEmitter.js
@@ -1,5 +1,3 @@
 import { EventEmitter as eventEmitter } from "events";
 
-// const eventEmitter = require("events");
-
 export const EventEmitter = new eventEmitter();
