+++ src/utils/misc/konami.ts
@@ -1,0 +1,26 @@
+import { useState } from "react";
+import useKonami from "react-use-konami";
+
+export const useKonamiEasterEgg = () => {
+  const [easterEgg, setEasterEgg] = useState(false);
+  const sixtyNine = () => {
+    setEasterEgg(true);
+  };
+
+  useKonami(sixtyNine, {
+    code: [
+      "ArrowUp",
+      "ArrowUp",
+      "ArrowDown",
+      "ArrowDown",
+      "ArrowLeft",
+      "ArrowRight",
+      "ArrowLeft",
+      "ArrowRight",
+      "b",
+      "a",
+    ],
+  });
+
+  return [easterEgg, setEasterEgg] as const;
+};
